Move methods to Pundit::Authorization module
The idea here is that the user should: `include Pundit::Authorization`
which only includes the necessary methods and classes.
